requirements.txt
google_adk
pymongo
pandas
matplotlib
sentence_transformers